Remove HAVE_S_ISREG
In NASM's configure.ac this checks whether S_ISREG exists as a function (also
see nasm/upstream/nasmlib/file.h) but POSIX says it's a macro and a source
code inspection confirms that that's also the case in glibc, FreeBSD, OpenBSD,
NetBSD, macOS, and Mingw-w64 so it's probably quite rare for it to be a
function. Therefore rather just let it remain #undef in the NASM build2
package.