Remove HAVE_STRING_H
Only referenced in NASM but not from any source code. Also seems like this
header should be present on any system that supports POSIX and/or some version
of standard C. I think HAVE_STRINGS_H (which applies to <strings.h> and
already exists in libbuild2-autoconf) is more likely to be used.