Remove HAVE_TYPEOF
This check does emulate NASM's custom check, but defining a `typeof` macro if
it's already a keyword (e.g., if in C23 or GNU modes) is not allowed by
C. Also not sure redefining it if it was a macro would be a good idea.

A general HAVE_TYPEOF check should probably do nothing but detect the presence
of the `typeof` keyword. But it looks like that would entail detecting
combinations of compiler and C mode/standard values. E.g. GCC or Clang in GNU
or C23 modes; or MSVC in C23 mode; etc.

We could add something like this but it seems like it could be tricky and it
wouldn't satisfy the requirements of NASM (i.e., the definition of a `typeof`
macro) which is the only consumer at this stage. So for the time being I've
added the check to the NASM project catalogue.

(For future reference, MSVC has had __typeof__ for a while and `typeof` was
added along with C23 support.)