Rename HAVE_STDNORETURN.h to HAVE_STDNORETURN_H.h
This check worked perfectly when it was still in the builtin catalog but NASM
would no longer build once the check was moved to its project catalog. The
error was "variable HAVE_STDNORETURN_H not defined". So looks like the build2
parser just ignored the macro when it was in the project catalog.